Use zExec.unsafeRunInterruptible where appropriate
As a fix for #1147 zExec.unsafeRunInterruptible was created, because when using
the pattern
  zExec.unsafeRun(<sometask>.uninterruptible)
it was possible to be interrupted before <sometask> was started causing
callbacks to be missed.

Replacing the above with zExec.unsafeRunInterruptible(<sometask>) fixed this.

This commit replaces the remaining instances of this pattern with
unsafeRunUninterruptible. At least once of these fixes a similar race condition
observed in #1252.