[#1813] Add QoL to NettyDriver ------------------------------
'NettyDriver' now supports a layer constructor that additionally requires Netty's 'EventLoopGroup' and 'ChannelFactory'.  This is layer provided via 'NettyDriver.manual'.

The reason for this addition is that under certain circumstances e.g. integration testing, it is beneficial to be able to share Netty dependencies across multiple Http app servers and test specs.

Housekeeping
------------

Remove 'settings.json', it shouldn't be here.