feat(endpoint): automatically validate body
Uses the `Schema.validate` method from `zio-schema` to validate the body
of a request based on the schema of the input type. If schema validation
fails, `HttpCodecError.InvalidEntity` is passed to the error channel,
indicating that the body is well-formed, but contains invalid values in
one or more fields.  A `wrap` method is included for `InvalidEntity`
which simplifies the process of converting `Schema.validate`'s
output (`Chunk[ValidationError]`) to `InvalidEntity` in a consistent
way.

resolves #2274