#2373
Whenever a ReadTimeoutException is thrown, the specific handling code within the case match will be executed. For now, I've left it to do nothing (i.e., it'll just swallow the exception). If you'd like to handle it in some other manner, like logging it at a different severity or executing some custom logic, replace the comment with the appropriate code.