Maintain query parameter ordering in QueryParams
The upstream Netty request uses a java.util.LinkedHashMap to store
the query string parameters. By using an ordered collection, it is
possible to exactly reconstruct the request url, which is sometimes
necessary to implement request signing and other security measures.

This change uses a scala.collection.immutable.SeqMap inside the
QueryParams to allow this ordering to be maintained while keeping
the convenience of the Map interface.