feat: Proper imports for generated openapi code
Extend the code generator, such that each generation step
can add imports, which are then rendered at the file or object level.

This fixes a few issues where imports were missing for generated case
classes, that had `Chunk` fields. It also unifies the handling of UUIDs
and opens the dor for additional future configuration options. For
example the collection type can be made configurable, or user defined
types can be added to the imports.

`writeFiles` now returns a list of written files. This allows for use
in sbt code generation tasks. I will add an sbt plugin in a later PR.