added alternative encoding of optional query parameters.
instead of `.query[Option[String]](query("a").optional)` the following can be used `.query[Option[String]](queryOptional("a"))`.

This approach does not double the amount of handlers as no fallback handler is needed.

Also added `queryAllOptional` that allows none or multiple query parameters. Until now `queryAll("x").optional` would be used, but the result type of the codec is a bit strange type `Option[Chunk[_]]`` instead of `Chunk[_]` that might be empty.