Make error handling safer.
This set of changes improves on the initial pass in a couple of ways.

There is no longer a need to await on the 'onFailure' promise, reasons provided below.

All non-websocket channel handling operations now live inside 'ClientInboundHandler'.  This includes the logic for 'exceptionCaught' which is responsible for properly completing 'onComplete', 'onResponse' and 'onFailure' when an exception occurs during Netty channel pipeline processing.

The 'ChannelInterface.interrupt' implementation now handles completion up of the 'onResponse' and 'onFailure' promises in the case where 'onComplete' is interrupted.

The close future listener has been completely removed as its responsibilities are covered by the 'ClientInboundHandler' and the 'ChannelInterface'. Interesting discovery for me was that the listener actually gets invoked BEFORE the 'exceptionCaught' method of 'ClientInboundHandler'.  This is what lead be to the decision to completely remove the listener all together.